hw/intc: Implment GIC-500
Implement GIC-500 from GICv3 family for arm64

This patch is a first step toward 128 cores support for arm64.

At first only 64 cores are supported.
This is because the largest integer type has the size of 64 bits and modifying
essential data structures in order to support 128 cores will require
the usage of bitops.

Currently SPI are not supported. This patch implements only what is necessary
in order to run the hw/arm/virt.c board with 64 cores.

Things left to do:

Add support to 128 cores.
Implement SPI.

Signed-off-by: Shlomo Pongratz <shlomo.pongratz@huawei.com>