Add virt-v3 machine that uses GIC-500
I would like to offer this, slightly improved implementation. The key thing is a new
kernel_irqchip_type member in Machine class. Currently it it used only by virt machine for
its internal purposes, however in future it is to be passed to KVM in
kvm_irqchip_create(). The variable is defined as int in order to be architecture agnostic,
for potential future users.

Signed-off-by: Pavel Fedin <p.fedin@samsung.com>